+++ src/onegov.election_day/onegov/election_day/tests/layouts/test_election_compound_layout.py
@@ -15,12 +15,14 @@ def test_election_compound_layout(session):
     compound = session.query(ElectionCompound).one()
 
     layout = ElectionCompoundLayout(compound, DummyRequest())
-    assert layout.all_tabs == ('districts', 'candidates', 'parties', 'data')
+    assert layout.all_tabs == (
+        'districts', 'candidates', 'party-strengths', 'data'
+    )
     assert layout.title() == ''
     assert layout.title('undefined') == ''
     assert layout.title('districts') == '__districts'
     assert layout.title('candidates') == 'Elected candidates'
-    assert layout.title('parties') == 'Parties'
+    assert layout.title('party-strengths') == 'Party strengths'
     assert layout.title('data') == 'Downloads'
     assert layout.main_view == 'ElectionCompound/districts'
     assert list(layout.menu) == []