+++ src/onegov.election_day/onegov/election_day/tests/views/test_views.py
@@ -178,7 +178,7 @@ def test_view_last_modified(election_day_app):
             '/election/election/candidates',
             '/election/election/statistics',
             '/election/election/panachage',
-            '/election/election/parties',
+            '/election/election/party-strengths',
             '/election/election/data',
             '/elections/elections',
             '/elections/elections/districts',
@@ -300,7 +300,7 @@ def test_view_svg(election_day_app):
         '/election/proporz-election/candidates-svg',
         '/election/proporz-election/panachage-svg',
         '/election/proporz-election/connections-svg',
-        '/election/proporz-election/parties-svg',
+        '/election/proporz-election/party-strengths-svg',
     )
     for path in paths:
         assert client.get(path, expect_errors=True).status_code == 503
@@ -350,7 +350,7 @@ def test_view_svg(election_day_app):
         'proporz-election-list-connections.svg',
         'proporz-election-lists.svg',
         'proporz-election-panachage.svg',
-        'proporz-election-parties.svg',
+        'proporz-election-party-strengths.svg',
         'vote-proposal.svg'
     ]