+++ src/onegov.election_day/onegov/election_day/tests/views/test_views_manage.py
@@ -192,7 +192,7 @@ def test_view_clear_results(election_day_app):
         '/election/proporz-election/lists',
         '/election/proporz-election/candidates',
         '/election/proporz-election/connections',
-        '/election/proporz-election/parties',
+        '/election/proporz-election/party-strengths',
         '/election/proporz-election/panachage',
         '/election/proporz-election/statistics',
         '/vote/vote'